backend/api/requirements.txt
flask
flask-restful
google-cloud-firestore
gunicorn